Refactor code to separate EOS functions into separate module, eos_functions.py. Modify tests accordingly. All tests passed.
